+++ modules/code-editor/src/views/full/utils/tree.ts
@@ -1,4 +1,4 @@
-import { ITreeNode } from '@blueprintjs/core'
+import { ITreeNode, Tree } from '@blueprintjs/core'
 import find from 'lodash/find'
 
 import { EditableFile } from '../../../backend/typings'
@@ -53,4 +53,58 @@ export const buildTree = (files: EditableFile[], expandedNodeIds: object) => {
   })
 
   return tree.childNodes
+}
+
+export const renameTreeNode = async (
+  node: ITreeNode<any>,
+  treeRef: React.RefObject<Tree<any>>,
+  callback: (newName: string) => Promise<void>
+) => {
+  const nodeDomElement = treeRef.current.getNodeContentElement(node.id)
+
+  const input = document.createElement('input')
+  input.type = 'text'
+  input.className = 'bp3-input bp3-small'
+  input.value = node.label as string
+
+  const div = document.createElement('div')
+  div.className = nodeDomElement.className
+
+  div.appendChild(input)
+
+  nodeDomElement.replaceWith(div)
+
+  input.focus()
+  input.select()
+
+  const closeRename = async e => {
+    e.preventDefault()
+    div.replaceWith(nodeDomElement)
+    window.removeEventListener('keydown', keyboardListener)
+    window.removeEventListener('mousedown', mouseListener)
+    const file = node.nodeData as EditableFile
+    let newName = input.value as string
+    newName = newName.endsWith('.js') ? newName : newName + '.js'
+    if (newName === node.label) {
+      return
+    }
+    try {
+      await callback(newName)
+    } catch (e) {
+      return
+    }
+    node.label = newName
+  }
+  const keyboardListener = async e => {
+    if (e.key === 'Enter' || e.key === 'Escape') {
+      await closeRename(e)
+    }
+  }
+  const mouseListener = async e => {
+    if (!div.contains(e.target)) {
+      await closeRename(e)
+    }
+  }
+  window.addEventListener('keydown', keyboardListener)
+  window.addEventListener('mousedown', mouseListener)
 }
